% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{datasim}
\alias{datasim}
\title{Generate simulated data}
\usage{
datasim(
  seed = NULL,
  p = 10,
  ni = 100,
  K = 10,
  error.Dis = "n",
  tau = 1/2,
  dgp = "dgp1",
  ...
)
}
\arguments{
\item{seed}{Random seed (optional)}

\item{p}{Variable dimension, default is 10}

\item{ni}{Sample size per group, default is 100}

\item{K}{Number of groups, default is 10}

\item{error.Dis}{Error distribution, default is 'n'}

\item{tau}{Quantile level, default is 1/2}

\item{dgp}{Data generating process, default is 'dgp1'}

\item{...}{Other parameters}
}
\value{
Returns a list containing simulated data including Y.hat (predicted values),
  X (independent variables), tB (regression coefficient matrix), Y.obs (observed values),
  and para.input (input parameters).
}
\description{
This function generates simulated data.
}
\examples{
dgp <- 'dgp1'
p <- 400
ni <- 20
K <- 10
error_dis <- 't'
tau <- 3/4
intercept <- !tau == 1/2
para.cases <- list(dgp = dgp, p = p, ni = ni, K = K, error.Dis = error_dis, tau = tau) # para for cases
data <- datasim(seed = 0, p = p, ni = ni, K = K, error.Dis = error_dis, tau = tau, dgp = dgp)

}
