% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIAQ.R
\name{RIAQ}
\alias{RIAQ}
\title{RIAQ: Robust Integrative Analysis via Quantile Regression with Homogeneity and Sparsity}
\usage{
RIAQ(
  X_temp,
  Y_temp,
  beta0 = NULL,
  tau,
  intercept = F,
  pen = "MCP",
  cc = 1,
  ini.lambda = 0.01,
  rho1 = 1,
  rho2 = 1,
  rho3 = 1,
  tol = 1e-04,
  nADMM = 500,
  is.show = T,
  para.turning = NULL,
  ...
)
}
\arguments{
\item{X_temp}{The input data matrix.}

\item{Y_temp}{The response variable.}

\item{beta0}{Initial value for beta (optional).}

\item{tau}{The quantile level.}

\item{intercept}{Logical value indicating whether to include an intercept term.}

\item{pen}{The penalty function type (e.g., 'MCP', 'SCAD').}

\item{cc}{A tuning parameter for the penalty function.}

\item{ini.lambda}{The initial value for the tuning parameter lambda.}

\item{rho1}{A tuning parameter.}

\item{rho2}{A tuning parameter.}

\item{rho3}{A tuning parameter.}

\item{tol}{Tolerance level for convergence.}

\item{nADMM}{Maximum number of ADMM iterations.}

\item{is.show}{Logical value indicating whether to display progress messages.}

\item{para.turning}{List of turning parameters (optional).}

\item{...}{Additional arguments.}
}
\value{
A list containing the estimated coefficients (Bhat).
}
\description{
Implementation of the method proposed in the paper "Robust Integrative Analysis via Quantile Regression with Homogeneity and Sparsity."
}
\examples{
\dontrun{
dgp = 'dgp1'
p = 400
ni = 20
K = 10
error_dis = 't'
tau = 3/4
intercept = !tau ==1/2

para.cases = list(dgp = dgp, p = p, ni = ni, K = K, error.Dis = error_dis, tau = tau) # para for cases
data = datasim(seed = 0, p = p, ni = ni, K = K, error.Dis = error_dis,tau = tau,dgp = dgp)

para.algo = list(rho1 = 1, rho2 = 1, tol = 1e-4, intercept = intercept , nADMM = 500, pen = 'MCP', cc = 1, ini.lambda = 0.005, is.show = T,para.turning = NULL) # paras for algo
args = c(para.cases, para.algo,list(X_temp = data$X, Y_temp = data$Y.obs, beta0 = t(data$tB)))
fit= do.call(RIAQ, args = args)

Beta_hat = fit$Bhat
}

}
\references{
Zeng, Hao, Chuang Wan, Wei Zhong, and Tuo Liu. 2023. “Robust Integrative Analysis via Quantile Regression with Homogeneity and Sparsity.” Working paper.
}
